IN-349 feat: put logic into a hook and add authorization checks and error handling
